Minor tweaks to the flex array

+++ include/hatrack/flexarray.h
@@ -77,7 +77,7 @@ void        *flexarray_get                (flexarray_t *, uint64_t, int *);
 bool         flexarray_set                (flexarray_t *, uint64_t, void *);
 void         flexarray_grow               (flexarray_t *, uint64_t);
 void         flexarray_shrink             (flexarray_t *, uint64_t);
-uint32_t     flexarray_len                (flexarray_t *);
+uint64_t     flexarray_len                (flexarray_t *);
 flex_view_t *flexarray_view               (flexarray_t *);
 void        *flexarray_view_next          (flex_view_t *, bool *);
 void         flexarray_view_delete        (flex_view_t *);
